refactor: apply one-line code quality improvements
- Replace chars().any(!is_ascii) with str::is_ascii()
- Replace manual ceiling division with u32::div_ceil()
- Remove redundant isize cast on hwnd.get()
- Remove redundant closure in and_then(hex_to_hsla)
- Remove needless ..Default::default() from TitlebarOptions literals
- Remove no-op .min(255) on u8 values in mouse encoding
- Use wrapping_add for explicit overflow semantics in UTF-8 mouse mode

+++ crates/codirigent-ui/src/workspace/render.rs
@@ -81,7 +81,7 @@ impl WorkspaceView {
         {
             use raw_window_handle::HasWindowHandle;
             let raw_handle = window.window_handle().ok().map(|h| match h.as_raw() {
-                raw_window_handle::RawWindowHandle::Win32(win32) => win32.hwnd.get() as isize,
+                raw_window_handle::RawWindowHandle::Win32(win32) => win32.hwnd.get(),
                 _ => 0,
             });
             if let Some(hwnd) = raw_handle {
